@@ -1,27 +1,23 @@
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 // Identifiers
 
-group = 'com.blacklocus'
+group = 'com.precioustatus'
 version = '0.3.5-SNAPSHOT'
-ext.isReleaseVersion = !version.endsWith("SNAPSHOT")
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 // Plugins
 
 apply plugin: 'java'
-sourceCompatibility = 1.6
-targetCompatibility = 1.6
-
+sourceCompatibility = 1.8
+targetCompatibility = 1.8
 apply plugin: 'maven'
-apply plugin: 'signing'
-apply plugin: 'idea'
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 // Defines
 
 /** Function always returns a new manifest that can be customized */
 def defaultManifest() {
-    return manifest{
+    return manifest {
         def git_cmd = "git rev-parse HEAD"
         def git_proc = git_cmd.execute()
         attributes 'SCM-Revision': git_proc.text.trim()
