@@ -1,12 +1,12 @@
 /**
  * Copyright 2013 BlackLocus
- *
+ * <p>
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -22,7 +22,6 @@ import com.codahale.metrics.Timer;
 import org.junit.Ignore;
 import org.junit.Test;
 
-import java.util.concurrent.Callable;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
@@ -39,83 +38,55 @@ public class CloudWatchReporterTest {
         ExecutorService executors = Executors.newCachedThreadPool();
 
         // increments the counter by 1 every second, meter ticked once, histogram ticked once, gauge given 1
-        executors.submit(new Callable<Void>() {
-            @Override
-            public Void call() throws Exception {
+        executors.submit(() -> {
+            MetricRegistry metricRegistry = new MetricRegistry();
+            new CloudWatchReporter(
+                    metricRegistry,
+                    CloudWatchReporterTest.class.getSimpleName(),
+                    new AmazonCloudWatchAsyncClient()
+            ).start(1, TimeUnit.MINUTES);
 
-                MetricRegistry metricRegistry = new MetricRegistry();
-                new CloudWatchReporter(
-                        metricRegistry,
-                        CloudWatchReporterTest.class.getSimpleName(),
-                        new AmazonCloudWatchAsyncClient()
-                ).start(1, TimeUnit.MINUTES);
+            metricRegistry.register("TheGauge", (Gauge<Long>) () -> 1L);
+            // Should be ignored by reporter
+            metricRegistry.register("TheGauge notNumeric", (Gauge<String>) () -> "yellow");
 
-                metricRegistry.register("TheGauge", new Gauge<Long>() {
-                    @Override
-                    public Long getValue() {
-                        return 1L;
-                    }
-                });
-                // Should be ignored by reporter
-                metricRegistry.register("TheGauge notNumeric", new Gauge<String>() {
-                    @Override
-                    public String getValue() {
-                        return "yellow";
-                    }
-                });
-
-                while(!Thread.interrupted()) {
-                    metricRegistry.counter("TheCounter TestDim=Yellow TestToken* machine=number1*").inc(1);
-                    metricRegistry.meter("TheMeter").mark();
-                    metricRegistry.histogram("TheHistogram").update(1);
-                    metricRegistry.histogram("TheHistogram").update(1);
-                    Timer.Context theTimer = metricRegistry.timer("TheTimer").time();
-                    Thread.sleep(1000);
-                    theTimer.close();
-                }
-                return null;
+            while (!Thread.interrupted()) {
+                metricRegistry.counter("TheCounter TestDim=Yellow TestToken* machine=number1*").inc(1);
+                metricRegistry.meter("TheMeter").mark();
+                metricRegistry.histogram("TheHistogram").update(1);
+                metricRegistry.histogram("TheHistogram").update(1);
+                Timer.Context theTimer = metricRegistry.timer("TheTimer").time();
+                Thread.sleep(1000);
+                theTimer.close();
             }
+            return null;
         });
 
         // increments the counter by 2 every second, meter ticked twice, histogram ticked twice, gauge given 2
-        executors.submit(new Callable<Void>() {
-            @Override
-            public Void call() throws Exception {
+        executors.submit(() -> {
+            MetricRegistry metricRegistry = new MetricRegistry();
+            new CloudWatchReporter(
+                    metricRegistry,
+                    CloudWatchReporterTest.class.getSimpleName(),
+                    new AmazonCloudWatchAsyncClient()
+            ).start(1, TimeUnit.MINUTES);
 
-                MetricRegistry metricRegistry = new MetricRegistry();
-                new CloudWatchReporter(
-                        metricRegistry,
-                        CloudWatchReporterTest.class.getSimpleName(),
-                        new AmazonCloudWatchAsyncClient()
-                ).start(1, TimeUnit.MINUTES);
+            metricRegistry.register("TheGauge", (Gauge<Long>) () -> 2L);
+            // Should be ignored by reporter
+            metricRegistry.register("TheGauge notNumeric", (Gauge<String>) () -> "green");
 
-                metricRegistry.register("TheGauge", new Gauge<Long>() {
-                    @Override
-                    public Long getValue() {
-                        return 2L;
-                    }
-                });
-                // Should be ignored by reporter
-                metricRegistry.register("TheGauge notNumeric", new Gauge<String>() {
-                    @Override
-                    public String getValue() {
-                        return "green";
-                    }
-                });
-
-                while(!Thread.interrupted()) {
-                    metricRegistry.counter("TheCounter TestDim=Yellow").inc(2);
-                    metricRegistry.meter("TheMeter").mark(3);
-                    metricRegistry.histogram("TheHistogram").update(2);
-                    Timer.Context theTimer = metricRegistry.timer("TheTimer").time();
-                    Thread.sleep(500);
-                    theTimer.close();
-                    theTimer = metricRegistry.timer("TheTimer").time();
-                    Thread.sleep(500);
-                    theTimer.close();
-                }
-                return null;
+            while (!Thread.interrupted()) {
+                metricRegistry.counter("TheCounter TestDim=Yellow").inc(2);
+                metricRegistry.meter("TheMeter").mark(3);
+                metricRegistry.histogram("TheHistogram").update(2);
+                Timer.Context theTimer = metricRegistry.timer("TheTimer").time();
+                Thread.sleep(500);
+                theTimer.close();
+                theTimer = metricRegistry.timer("TheTimer").time();
+                Thread.sleep(500);
+                theTimer.close();
             }
+            return null;
         });
 
         for (int i = 0; i < 60 /* one hour */; i++) {
