@@ -1,12 +1,12 @@
 /**
  * Copyright 2013 BlackLocus
- *
+ * <p>
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -17,7 +17,6 @@ package com.blacklocus.metrics;
 
 import com.amazonaws.services.cloudwatch.model.Dimension;
 import com.amazonaws.services.cloudwatch.model.MetricDatum;
-import com.google.common.base.Function;
 import com.google.common.base.Functions;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
@@ -31,17 +30,13 @@ import java.util.Set;
 /**
  * @author Jason Dunkelberger (dirkraft)
  */
+@SuppressWarnings("unchecked")
 public class DemuxedKeyTest {
 
     @Test
     public void testSingleName() {
         DemuxedKey key = new DemuxedKey("SingleToken");
-        List<MetricDatum> data = Lists.newArrayList(key.newDatums("test", new Function<MetricDatum, MetricDatum>() {
-            @Override
-            public MetricDatum apply(MetricDatum datum) {
-                return datum.withValue(3.5);
-            }
-        }));
+        List<MetricDatum> data = Lists.newArrayList(key.newDatums("test", datum -> datum.withValue(3.5)));
         Assert.assertEquals(1, data.size());
     }
 
@@ -118,7 +113,6 @@ public class DemuxedKeyTest {
     }
 
     @Test
-    @SuppressWarnings("unchecked")
     public void testMultiPermDimensions() {
         DemuxedKey key = new DemuxedKey("Name key=value* color=green* machine=localhost");
         List<MetricDatum> data = Lists.newArrayList(key.newDatums("testMultiPermDimensions", Functions.<MetricDatum>identity()));
@@ -133,26 +127,16 @@ public class DemuxedKeyTest {
     }
 
     boolean containsExactly(List<MetricDatum> data, String... names) {
-        return Sets.symmetricDifference(Sets.newHashSet(Lists.transform(data, new Function<MetricDatum, String>() {
-            @Override
-            public String apply(MetricDatum input) {
-                return input.getMetricName();
-            }
-        })), Sets.newHashSet(names)).isEmpty();
+        return Sets.symmetricDifference(Sets.newHashSet(Lists.transform(data, input -> input.getMetricName())), Sets.newHashSet(names)).isEmpty();
     }
 
     boolean containsExactly(List<MetricDatum> data, Set<Dimension>... dimensions) {
-        return Sets.newHashSet(Lists.transform(data, new Function<MetricDatum, Set<Dimension>>() {
-            @Override
-            public Set<Dimension> apply(MetricDatum input) {
-                return Sets.newHashSet(input.getDimensions());
-            }
-        })).equals(Sets.newHashSet(dimensions));
+        return Sets.newHashSet(Lists.transform(data, input -> Sets.newHashSet(input.getDimensions()))).equals(Sets.newHashSet(dimensions));
     }
 
     Set<Dimension> d(String... keyValuePairs) {
         Set<Dimension> dimensions = new HashSet<Dimension>(keyValuePairs.length / 2);
-        for (int i = 0; i < keyValuePairs.length; i+= 2) {
+        for (int i = 0; i < keyValuePairs.length; i += 2) {
             dimensions.add(new Dimension().withName(keyValuePairs[i]).withValue(keyValuePairs[i + 1]));
         }
         return dimensions;
